move evaluator code to its own package
